Support accumulate for linecharts
